db: update block fetching, fix typing
