+++ storage/storage.go
@@ -11,5 +11,6 @@ type Storage interface {
 	LastBlock(chainID int) (types.Block, error)
 	BlockCount(chainID int) (int, error)
 	UpsertBlock(block types.Block, chainID int) error
-	BlockAtHeight(height int, chainID int) (types.Block, error)
+	BlockAtHeight(height int64, chainID int) (types.Block, error)
+	GetUnlinkedChainTips(chainID int) ([]types.ChainTip, error)
 }
